evaluation/gold-comparison/requirements.txt
anthropic>=0.49.0
pyyaml>=6.0.2
pytest>=8.3.5
